refactor: 매 라운드마다 자동차들의 포지션 출력에 대한 책임을 OutputView에게 전가
